requirements.txt
langchain
google-generativeai
langchain-google-genai
python-dotenv